refactor: api 폴더 내부 파일 구조화
