+++ src/api/user-service.js
@@ -1,25 +1,14 @@
-// src/api/myPage.js
+// src/api/user-service.js
 import client from './client';
 
-/** 마이페이지 - 내 정보 조회 */
+// 마이페이지 사용자 정보 조회
 export function fetchMyInfo(accessToken) {
   return client.get('/user-service/mypage/info', {
     headers: { Authorization: `Bearer ${accessToken}` },
   });
 }
 
-/** ✅ 마이페이지 - 현재 주소/세대번호 조회 (GET) */
-export function fetchCurrentAddress(accessToken) {
-  return client.get('/user-service/current-address-and-number', {
-    headers: { Authorization: `Bearer ${accessToken}` },
-  });
-}
-
-/**
- * 마이페이지 - 내 정보 업데이트 (POST, multipart)
- * - JSON은 key=data (@RequestPart("data"))
- * - 이미지 파일은 key=image (@RequestPart("image"))
- */
+// 마이 페이지 사용자 정보 편집
 export function updateMyInfo(accessToken, payload = {}, imageFile = null) {
   const fd = new FormData();
   fd.append('data', new Blob([JSON.stringify(payload)], { type: 'application/json' }));
@@ -33,7 +22,14 @@ export function updateMyInfo(accessToken, payload = {}, imageFile = null) {
   });
 }
 
-/** 마이페이지 - 현재 주소 변경 (PATCH) */
+// 사용자 현재 주소 조회
+export function fetchCurrentAddress(accessToken) {
+  return client.get('/user-service/get/current-address-and-number', {
+    headers: { Authorization: `Bearer ${accessToken}` },
+  });
+}
+
+// 사용자 현재 주소 조회
 export function changeCurrentAddress(accessToken, payload) {
   return client.patch('/user-service/change-currentAddress', payload, {
     headers: { Authorization: `Bearer ${accessToken}` },
